pyproject.toml
[tool.poetry]
name = "warehouse"
version = "0.1.0"
description = ""
authors = ["coderhanzo <reinerbielert@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.2"
django = "^5.0.6"
django-environ = "^0.11.2"
mysqlclient = "^2.2.4"
mysql-connector-python = "^9.0.0"
django-phonenumber-field = "^8.0.0"
phonenumbers = "^8.13.40"
djoser = "^2.2.3"
django-mptt = "^0.16.0"
django-autoslug = "^1.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
